src/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
python-decouple==3.8
Werkzeug==3.0.3
WTForms==3.1.2
